+++ client/src/components/Card.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import CardAmpliada from "./CardAmpliada";
 import soldout from "../assets/img/soldOut.png";
 import { useNavigate } from "react-router-dom";
@@ -17,7 +17,6 @@ const Card = ({
   const [showAlert, setShowAlert] = useState(false);
   const [currentImage, setCurrentImage] = useState(imagen);
   const [showModal, setShowModal] = useState(false);
-
   const [selectedTalles, setSelectedTalles] = useState([]);
 
   const handleTalleChange = (talle) => {
@@ -29,6 +28,18 @@ const Card = ({
   };
 
   const navigate = useNavigate();
+
+  useEffect(() => {
+    if (showAlert) {
+      actualizarContadorCarrito();
+
+      const timeoutId = setTimeout(() => {
+        setShowAlert(false);
+      }, 2000);
+
+      return () => clearTimeout(timeoutId);
+    }
+  }, [showAlert, actualizarContadorCarrito]);
 
   const handleDetalleClick = () => {
     navigate(`/productos/${id}`);
@@ -76,10 +87,6 @@ const Card = ({
     localStorage.setItem("carrito", JSON.stringify(carrito));
 
     setShowAlert(true);
-
-    setTimeout(() => {
-      setShowAlert(false);
-    }, 2000);
     actualizarContadorCarrito();
   };
