+++ client/src/components/Detalle.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import Carousel from "react-multi-carousel";
 import "react-multi-carousel/lib/styles.css";
+import Layout from "../pages/Layout";
 
 export default function Detalle() {
   const [producto, setProducto] = useState(null);
@@ -40,74 +41,76 @@ export default function Detalle() {
   }, [id]);
 
   return (
-    <div className="DetalleCardContainer">
-      {producto ? (
-        <div className="DetalleCard">
-          <h2>{producto.nombre}</h2>
-          <Carousel
-            additionalTransfrom={0}
-            arrows
-            autoPlaySpeed={3000}
-            centerMode={false}
-            className=""
-            containerClass="container"
-            dotListClass=""
-            draggable
-            focusOnSelect={false}
-            infinite
-            itemClass=""
-            keyBoardControl
-            minimumTouchDrag={80}
-            pauseOnHover
-            renderArrowsWhenDisabled={false}
-            renderButtonGroupOutside={false}
-            renderDotsOutside={false}
-            responsive={{
-              desktop: {
-                breakpoint: {
-                  max: 3000,
-                  min: 1024,
+    <Layout>
+      <div className="DetalleCardContainer">
+        {producto ? (
+          <div className="DetalleCard">
+            <h2>{producto.nombre}</h2>
+            <Carousel
+              additionalTransfrom={0}
+              arrows
+              autoPlaySpeed={3000}
+              centerMode={false}
+              className=""
+              containerClass="container"
+              dotListClass=""
+              draggable
+              focusOnSelect={false}
+              infinite
+              itemClass=""
+              keyBoardControl
+              minimumTouchDrag={80}
+              pauseOnHover
+              renderArrowsWhenDisabled={false}
+              renderButtonGroupOutside={false}
+              renderDotsOutside={false}
+              responsive={{
+                desktop: {
+                  breakpoint: {
+                    max: 3000,
+                    min: 1024,
+                  },
+                  items: 1,
                 },
-                items: 1,
-              },
-              mobile: {
-                breakpoint: {
-                  max: 464,
-                  min: 0,
+                mobile: {
+                  breakpoint: {
+                    max: 464,
+                    min: 0,
+                  },
+                  items: 1,
                 },
-                items: 1,
-              },
-              tablet: {
-                breakpoint: {
-                  max: 1024,
-                  min: 464,
+                tablet: {
+                  breakpoint: {
+                    max: 1024,
+                    min: 464,
+                  },
+                  items: 1,
                 },
-                items: 1,
-              },
-            }}
-            rewind={false}
-            rewindWithAnimation={false}
-            rtl={false}
-            shouldResetAutoplay
-            showDots
-            sliderClass=""
-            slidesToSlide={1}
-            swipeable
-          >
-            <img src={producto.imagen} className="CarrouselImg" alt="" />
-            <img src={producto.imagen1} className="CarrouselImg" alt="" />
-            <img src={producto.imagen2} className="CarrouselImg" alt="" />
-            <img src={producto.imagen3} className="CarrouselImg" alt="" />
-          </Carousel>
+              }}
+              rewind={false}
+              rewindWithAnimation={false}
+              rtl={false}
+              shouldResetAutoplay
+              showDots
+              sliderClass=""
+              slidesToSlide={1}
+              swipeable
+            >
+              <img src={producto.imagen} className="CarrouselImg" alt="" />
+              <img src={producto.imagen1} className="CarrouselImg" alt="" />
+              <img src={producto.imagen2} className="CarrouselImg" alt="" />
+              <img src={producto.imagen3} className="CarrouselImg" alt="" />
+            </Carousel>
 
-          <p>{producto.marca}</p>
-          <p>Precio: $ {producto.precio}</p>
-          <p>Código: {producto.codigo}</p>
-          <p>Talles: {producto.talle}</p>
-        </div>
-      ) : (
-        <div>Cargando...</div>
-      )}
-    </div>
+            <p>{producto.marca}</p>
+            <p>Precio: $ {producto.precio}</p>
+            <p>Código: {producto.codigo}</p>
+            <p>Talles: {producto.talle}</p>
+          </div>
+        ) : (
+          <div>Cargando...</div>
+        )}
+      </div>
+    </Layout>
   );
 }
