+++ client/src/components/NavBarAlternativo.jsx
@@ -1,21 +1,70 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import logo from "../assets/img/logos/LOGO 1.jpeg";
 import { FiSearch, FiShoppingCart } from "react-icons/fi";
 import { FaUser } from "react-icons/fa";
 import useAuth from "../hooks/useAuth";
 import { Link } from "react-router-dom";
+import CarritoModal from "./CarritoModal";
+import Login from "./usuario/Login";
+import EditarUsuario from "./usuario/EditarUsuario";
 
-export default function NavBarAlternativo({
-  handleMostrarModalCarrito,
-  carritoC,
-  handleSearchByMarca,
-  handleMostrarModalLogin,
-  handleMostrarModalEdit,
-}) {
+export default function NavBarAlternativo({ onSearchByMarca }) {
   const [searchTerm, setSearchTerm] = useState("");
   const [suggestions, setSuggestions] = useState([]);
   const [showUserMenu, setShowUserMenu] = useState(false);
+  const [modalCarrito, setModalCarrito] = useState(false);
+  const [carritoC, setCarritoC] = useState(0);
   const { auth, setAuth } = useAuth();
+  const [edit, setEdit] = useState(false);
+  const [login, setLogin] = useState(false);
+
+  useEffect(() => {
+    const storedAuth = JSON.parse(localStorage.getItem("auth"));
+    if (storedAuth) {
+      setAuth(storedAuth);
+    }
+  }, [setAuth]);
+
+  useEffect(() => {
+    actualizarContadorCarrito();
+  }, [modalCarrito]);
+
+  const carrito = JSON.parse(localStorage.getItem("carrito")) || [];
+
+  const actualizarContadorCarrito = () => {
+    const count = carrito.length;
+    setCarritoC(count);
+  };
+
+  const handleMostrarModalCarrito = () => {
+    setModalCarrito(true);
+    actualizarContadorCarrito();
+  };
+
+  const handleCerrarModalCarrito = () => {
+    setModalCarrito(false);
+    actualizarContadorCarrito();
+  };
+
+  const handleMostrarModalLogin = () => {
+    setLogin(true);
+  };
+
+  const handleCerrarModalLogin = () => {
+    setLogin(false);
+  };
+
+  const handleMostrarModalEdit = () => {
+    setEdit(true);
+  };
+
+  const handleCerrarModalEdit = () => {
+    setEdit(false);
+  };
+
+  useEffect(() => {
+    actualizarContadorCarrito();
+  }, []);
 
   const handleInputChange = (event) => {
     const value = event.target.value;
@@ -31,7 +80,7 @@ export default function NavBarAlternativo({
 
   const handleSuggestionClick = (suggestion) => {
     setSearchTerm(suggestion);
-    handleSearchByMarca(suggestion);
+    onSearchByMarca(suggestion);
     setSuggestions([]);
   };
 
@@ -63,7 +112,7 @@ export default function NavBarAlternativo({
   };
 
   const handleSearch = () => {
-    handleSearchByMarca(searchTerm);
+    onSearchByMarca(searchTerm);
   };
 
   const handleKeyPress = (event) => {
@@ -104,6 +153,31 @@ export default function NavBarAlternativo({
 
   return (
     <div>
+      {modalCarrito && (
+        <div>
+          <CarritoModal
+            handleCerrarModalCarrito={handleCerrarModalCarrito}
+            actualizarContadorCarrito={actualizarContadorCarrito}
+          />
+        </div>
+      )}
+
+      {login && (
+        <div>
+          <Login
+            handleMostrarModalLogin={handleMostrarModalLogin}
+            handleCerrarModalLogin={handleCerrarModalLogin}
+          />
+        </div>
+      )}
+      {edit && (
+        <div>
+          <EditarUsuario
+            handleMostrarModalEdit={handleMostrarModalEdit}
+            handleCerrarModalEdit={handleCerrarModalEdit}
+          />
+        </div>
+      )}
       <div className="navbar-container">
         <div className="black-bar">
           <p>Prendas 100% ORIGINALES</p>
