+++ client/src/pages/FooterAlternativo.jsx
@@ -1,10 +1,28 @@
-import React from "react";
+import React, { useState } from "react";
 import mercadoPago from "../assets/img/logos/mercadopago.png";
+import AboutUs from "../components/AboutUs";
+import Contact from "../components/Contact";
 
-export default function FooterAlternativo({
-  handleMostrarModalContact,
-  handleMostrarModalAbout,
-}) {
+export default function FooterAlternativo() {
+  const [contact, setContact] = useState(false);
+  const [modal, setModal] = useState(false);
+
+  const handleMostrarModalAbout = () => {
+    setModal(true);
+  };
+
+  const handleCerrarModalAbout = () => {
+    setModal(false);
+  };
+
+  const handleMostrarModalContact = () => {
+    setContact(true);
+  };
+
+  const handleCerrarModalContact = () => {
+    setContact(false);
+  };
+
   return (
     <div className="principal-footer">
       <div className="footerAlternativo">
@@ -32,6 +50,24 @@ export default function FooterAlternativo({
         <img src={mercadoPago} alt="mp" />
         <h4 style={{ color: "black" }}>Transferencia / Depósito bancario</h4>
       </div>
+      {contact && (
+        <div className="modal">
+          {/* <div className="modal-content"> */}
+          <Contact handleCerrarModalContact={handleCerrarModalContact} />
+          {/* </div> */}
+        </div>
+      )}
+      {modal && (
+        <div className="modal">
+          {/* <div className="modal-content"> */}
+          <AboutUs handleCerrarModalAbout={handleCerrarModalAbout} />
+          {/* </div> */}
+        </div>
+      )}
+      <hr style={{ color: "grey", maxWidth: "100%" }} />
+      <span className="copy">
+        Copyright © 2024 | Venice Indumentaria Todos los derechos reservados
+      </span>
     </div>
   );
 }
