+++ client/src/pages/Layout.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import NavBarAlternativo from "../components/NavBarAlternativo";
+import FooterAlternativo from "./FooterAlternativo";
+
+const Layout = ({ children, onSearchByMarca }) => {
+  return (
+    <div>
+      <NavBarAlternativo onSearchByMarca={onSearchByMarca} />
+      {children}
+      <FooterAlternativo />
+    </div>
+  );
+};
+
+export default Layout;
